Reorganise to reuse dt protos
